backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
python-dotenv
faiss-cpu
PyPDF2
bcrypt==4.3.0

